tsl support

+++ src/main/java/com/authlete/common/dto/TslConfigData.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 import java.net.URI;
 
 /**
- * A class that represents trust status list configurations data
+ * A class that represents TSL configurations data
  * The set consists of the following.
  *
  * <ul>
@@ -17,48 +17,54 @@ import java.net.URI;
  * <li>{@code publishEndpoint}
  * </ul>
  *
- * @since TODO
- * @since Authlete TODO
+ * @since 4.30
+ * @since Authlete 3.1
  *
- * @see <a href="https://datatracker.ietf.org/doc/draft-ietf-oauth-status-list/12/"
+ * @see <a href="https://datatracker.ietf.org/doc/draft-ietf-oauth-status-list/13/"
  *      >Token Status List (TSL)</a>
  */
-public class TslConfigData implements Serializable {
+public class TslConfigData implements Serializable
+{
 
     private static final long serialVersionUID = 1L;
 
     /**
      * The format of the TSL. Possible values are JWT and CWT. Currently only JWT is supported
      *
-     * @since TODO
+     * @since 4.30
+     * @since Authlete 3.1
      */
     private TslPublishFormat format;
 
     /**
      * The validity of the TSL in hours
      *
-     * @since TODO
+     * @since 4.30
+     * @since Authlete 3.1
      */
-    private int validity;
+    private long validity;
 
     /**
      * Publish TSL after every X hours
      *
-     * @since TODO
+     * @since 4.30
+     * @since Authlete 3.1
      */
-    private int publishFrequency;
+    private long publishFrequency;
 
     /**
      * Time to live in hours which provides indication to verifiers to cache this TSL
      *
-     * @since TODO
+     * @since 4.30
+     * @since Authlete 3.1
      */
-    private int timeToLive;
+    private long timeToLive;
 
     /**
      * Endpoint where to publish this TSL
      *
-     * @since TODO
+     * @since 4.30
+     * @since Authlete 3.1
      */
     private URI publishEndpoint;
 
@@ -70,12 +76,11 @@ public class TslConfigData implements Serializable {
 
     /**
      * Copy constructor.
-     *
-     * @param tslConfigData
-     *         Source to copy data from. {@code null} won't raise any exception.
      */
-    public TslConfigData(TslConfigData tslConfigData) {
-        if (tslConfigData == null) {
+    public TslConfigData(TslConfigData tslConfigData)
+    {
+        if (tslConfigData == null)
+        {
             return;
         }
         format = tslConfigData.getFormat();
@@ -85,49 +90,150 @@ public class TslConfigData implements Serializable {
         publishEndpoint = tslConfigData.getPublishEndpoint();
     }
 
-    public TslConfigData setFormat(TslPublishFormat format) {
+    /**
+     * Sets the TSL publish format.
+     *
+     * @param format the {@link TslPublishFormat} to use for publishing.
+     *
+     * @return this {@link TslConfigData} instance for method chaining
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public TslConfigData setFormat(TslPublishFormat format)
+    {
         this.format = format;
 
         return this;
     }
-    public TslPublishFormat getFormat() {
+
+    /**
+     * Returns the configured TSL publish format.
+     *
+     * @return the {@link TslPublishFormat} in use.
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public TslPublishFormat getFormat()
+    {
         return format;
     }
 
-    public TslConfigData setValidity(int validity) {
+    /**
+     * Sets the validity period for the TSL in hours.
+     *
+     * @param validity the validity duration
+     * @return this {@link TslConfigData} instance for method chaining
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public TslConfigData setValidity(long validity) {
         this.validity = validity;
 
         return this;
     }
-    public int getValidity() {
+
+    /**
+     * Returns the validity period of the TSL.
+     *
+     * @return the validity duration
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public long getValidity()
+    {
         return validity;
     }
 
-    public TslConfigData setPublishFrequency(int publishFrequency) {
+    /**
+     * Sets the frequency in hours at which the TSL should be published.
+     *
+     * @param publishFrequency the publishing frequency
+     * @return this {@link TslConfigData} instance for method chaining
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     *
+     */
+    public TslConfigData setPublishFrequency(long publishFrequency)
+    {
         this.publishFrequency = publishFrequency;
 
         return this;
     }
-    public int getPublishFrequency() {
+
+    /**
+     * Returns the publish frequency of the TSL.
+     *
+     * @return the publishing frequency
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public long getPublishFrequency()
+    {
         return publishFrequency;
     }
 
-    public TslConfigData setTimeToLive(int timeToLive) {
+    /**
+     * Sets the time to live in hours which provides indication to verifiers to cache this TSL
+     *
+     * @param timeToLive the TTL value
+     * @return this {@link TslConfigData} instance for method chaining
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public TslConfigData setTimeToLive(long timeToLive)
+    {
         this.timeToLive = timeToLive;
 
         return this;
     }
-    public int getTimeToLive() {
+
+    /**
+     * Returns the time-to-live (TTL) in hours value for the TSL.
+     *
+     * @return the TTL value
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public long getTimeToLive()
+    {
         return timeToLive;
     }
 
-    public TslConfigData setPublishEndpoint(URI publishEndpoint) {
+    /**
+     * Sets the endpoint URI where the TSL will be published.
+     *
+     * @param publishEndpoint the publish endpoint URI.
+     *
+     * @return this {@link TslConfigData} instance for method chaining
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public TslConfigData setPublishEndpoint(URI publishEndpoint)
+    {
         this.publishEndpoint = publishEndpoint;
 
         return this;
     }
-    public URI getPublishEndpoint() {
+
+    /**
+     * Returns the endpoint URI where the TSL is published.
+     *
+     * @return the publish endpoint URI.
+     *
+     * @since 4.30
+     * @since Authlete 3.1
+     */
+    public URI getPublishEndpoint()
+    {
         return publishEndpoint;
     }
 }
-
